+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   - com.dwarfeng.familyhelper.life.stack.service.ActivityTemplateCoverOperateService。
   - com.dwarfeng.familyhelper.life.stack.service.ActivityTemplateParticipantOperateService。
   - com.dwarfeng.familyhelper.life.stack.service.ActivityTemplateFileOperateService。
+  - com.dwarfeng.familyhelper.life.stack.service.ActivityTemplateDataInfoOperateService。
 
 - 新增实体及其维护服务，单元测试通过。
   - com.dwarfeng.familyhelper.life.stack.bean.entity.Activity。